Cleanup søskenjustering
